feat(GetSubscriptionsFromUserHandler)

+++ src/site/Bl.FeatureFlag.Application/Commands/GetSubscriptionsFromUser/GetSubscriptionsFromUserRequest.cs
@@ -2,6 +2,6 @@
 
 public record GetSubscriptionsFromUserRequest(
     Guid UserId,
-    long Skip,
-    long Take)
+    int Skip,
+    int Take)
     : IRequest<GetSubscriptionsFromUserResponse>;
